[Feat/#25] Spring Batch 를 이용한 2만5천개의 멤버, 30만개의 숙소, 300만건의 예약 더미 데이터 생성 (#28)

+++ src/main/java/com/airdnb/clone/domain/stay/entity/StayFee.java
@@ -10,6 +10,7 @@ import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.ToString;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -20,10 +21,10 @@ import lombok.NoArgsConstructor;
 public class StayFee {
 
     @Column(name = "PER_NIGHT")
-    private Integer perNight;
+    private Long perNight;
 
     @Column(name = "CLEANING_FEE")
-    private Integer cleaningFee;
+    private Long cleaningFee;
 
     public Long calculateTotalRate(LocalDateTime checkIn, LocalDateTime checkOut) {
         validate(checkIn, checkOut);
